minor tweaks
added a loadbang to the ASR of the Dynamic ocean sounds, because they weren't playing without setting an init value. added another loadbang to the [lop~] for the Static ocean sounds so it had a lowpass filter from the get go. Also switched the noise~ from the ocean sounds to receive noise from [s~ noiseSend] to save cpu power.